Extract workspace folder initialization
