+++ src/__tests__/server/initialization.test.ts
@@ -1,0 +1,16 @@
+import { describe, expect, it } from "bun:test";
+import { workspaceFoldersFromInitialize } from "../../server/initialization";
+
+describe("server initialization", () => {
+  it("uses rootUri when workspace folders are unavailable", () => {
+    expect(workspaceFoldersFromInitialize({
+      workspaceFolders: null,
+      rootUri: "file:///workspace",
+    })).toEqual([
+      {
+        uri: "file:///workspace",
+        name: "file:///workspace",
+      },
+    ]);
+  });
+});
